Add and modify earthly lint targets to match the workflow
In case someone want to run them locally before pushing to CI.
This beats the purpose of having a re-usable workflow but we don't care
much if the targets diverge from the workflow. They are only meant to
save us some CI cycles by running the checks locally before pushing. If
they diverge, worst case scenario, we'll see some failures on CI that
need to be fixed.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>