Drop upx compression
golang official image no longer has the upx binary available. If we want
to keep on the latests we should not fix the image to bullseye and keep
on the current and future version tag for golang.

This patch drops upx and just uses the golang image for golang deps

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>