Create empty k3s service file
Somehow on systemd 245 (245.4-4ubuntu3.22) becuase the k3s file was
created with no content, it marked it as a masked service and nothing
would work as expected.

Creating the file with empty service and execstart makes the service not
appear as masked and allows the override to work as expected

This patch also adds more logging to the log file, and writing the
service file before starting it.

```
root@qcs6490-odk:~# systemctl status k3s
Warning: The unit file, source configuration file or drop-ins of k3s.service changed on disk. Run 'systemctl daemon-reload' to reload units.
● k3s.service
     Loaded: masked (Reason: Unit k3s.service is masked.)
    Drop-In: /etc/systemd/system/k3s.service.d
             └─override.conf
     Active: inactive (dead)
```

Signed-off-by: Itxaka <itxaka@kairos.io>