Remove the bus.EventAvailableReleases implementation
to let the kairos-agent handle upgrades from now on.
It used to be that provider-kairos returned releases from the
provider-kairos repository but it's no longer the case. Everything is
release in the kairos-io/kairos repository and the new versioneer
library of the kairos-sdk (which is used in the kairos-agent) is
capabable of handling both "standard" and "core" images.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>