Remove the bus.EventAvailableReleases implementation (#496)
to let the kairos-agent handle upgrades from now on. It used to be that
provider-kairos returned releases from the provider-kairos repository
but it's no longer the case. Everything is release in the
kairos-io/kairos repository and the new versioneer library of the
kairos-sdk (which is used in the kairos-agent) is capabable of handling
both "standard" and "core" images.

Part of: https://github.com/kairos-io/kairos/issues/1999

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>