feat: implement build and info event handlers for provider
Add BuildEvent and InfoEvent functions to handle buildtime and info events for the provider. The BuildEvent function downloads and installs the specified provider (k3s or k0s) based on the event payload, while the InfoEvent retrieves the installed version of the provider. This enhances the provider's functionality and integration with kairos-init.

Signed-off-by: Itxaka <itxaka@kairos.io>